Implemented: Feature Request: external tool: multi instances     fixes: https://github.com/eranif/codelite/issues/1016

+++ ExternalTools/externaltoolsdata.cpp
@@ -79,7 +79,7 @@ void ExternalToolsData::DeSerialize(Archive& arch)
     m_tools.clear();
     for(size_t i = 0; i < count; i++) {
         ToolInfo info;
-        arch.Read(wxString::Format(wxT("Tool_%lu"), i), &info);
+        arch.Read(wxString::Format(wxT("Tool_%d"), (int)i), &info);
         m_tools.push_back(info);
     }
 }
@@ -91,6 +91,6 @@ void ExternalToolsData::Serialize(Archive& arch)
 
     // now loop over the toolsinfo and serialize them
     for(size_t i = 0; i < m_tools.size(); i++) {
-        arch.Write(wxString::Format(wxT("Tool_%lu"), i), &m_tools.at(i));
+        arch.Write(wxString::Format(wxT("Tool_%d"), (int)i), &m_tools.at(i));
     }
 }
